start variadic function

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -12,14 +12,14 @@
 int sum_them_all(const unsigned int n, ...)
 {
 	va_list ap;
-	int i, sum;
+	unsigned int i, sum;
 
 	if (n == 0)
 		return (0);
 	va_start(ap, n);
 	sum = 0;
-	for (i = 0; i < (int)n; i++)
-		sum += va_arg(ap, int); 
+	for (i = 0; i < n; i++)
+		sum += va_arg(ap, int);
 	va_end(ap);
 	return (sum);
 }
